Sprint 3 4:25PM 10/30/25
While designing the checkSOSFormed function, I ran into some logical difficulties with how to apply it to the subclasses simpleSOSGame and generalSOSGame, which I haven't fully built yet. I took some time to research how to make my SOSGame class more generalized in a way that it could be most efficiently extended to the simple and general subclasses, as well as asked DeepSeek LLM for some assistance in generalizing my SOSGame class. I am enacting changes to the structure of the SOSGame class, and will make regular pushes as I update the class with more details.