Sprint 3 4:42PM 10/30/25
Added functionality to checkSOSFormed to call the incrementScore() function and the drawSOSChain() function, so if an SOS or multiple SOS are formed, they will be drawn on the board, and the player's score will be incremented. In the simple game, this will end the game, in the general game, this will increment the player's score for the appropriate amount of points correlating to the SOSs created.